+++ packages/react/src/index.ts
@@ -30,6 +30,14 @@ export { default as GridBase } from "./atoms/grid/base";
 export { default as TextInputBase } from "./atoms/textInputs/base";
 export { default as PasswordTextInputBase } from "./atoms/textInputs/password";
 
+// Checkbox
+
+export { default as CheckboxBase } from "./atoms/checkbox/base";
+
+// Toggle
+
+export { default as ToggleBase } from "./atoms/toggles/base";
+
 // Tags
 
 export { default as TagInputBase } from "./molecules/tags/base";
@@ -84,6 +92,7 @@ export { default as ModalConfirm } from "./molecules/modals/confirm";
 
 export { default as DropdownBase } from "./molecules/dropdowns/base";
 export { default as DropdownMulti } from "./molecules/dropdowns/multi";
+export { default as DropdownBrowser } from "./molecules/dropdowns/browser";
 
 // Navigations