+++ packages/react/src/organisms/navigations/base.tsx
@@ -1,4 +1,4 @@
-import React, { ReactNode } from "react";
+import React, { ReactNode, useState } from "react";
 import "../../../tailwind.css";
 import { useThemeContext } from "../../theme/ThemeProvider";
 import resolvedStyleProps from "../../utils/resolvedStyleProps";
@@ -8,6 +8,7 @@ export interface NavigationBaseProps {
   left?: any;
   middle?: any;
   right?: any;
+  sidebar?: any;
   leftArrangement?: string;
   middleArrangement?: string;
   rightArrangement?: string;
@@ -18,7 +19,8 @@ export interface NavigationBaseProps {
 }
 
 const NavigationBase = (props: NavigationBaseProps) => {
-  //   const theme: any = useThemeContext();
+  const theme: any = useThemeContext();
+  const [sidebar, setSidebar] = useState<boolean>(false);
 
   const backgroundClasses = props.backgroundClasses || "bg-gray-50";
   const borderClasses = props.borderClasses || "border-b border-gray-200";
@@ -26,6 +28,13 @@ const NavigationBase = (props: NavigationBaseProps) => {
     props.widthClasses || "mx-auto max-w-7xl px-4 sm:px-6 lg:px-8";
   const heightClasses =
     props.heightClasses || "flex h-16 items-center justify-between";
+
+  const toggleSidebar = () => {
+    console.log("toggleSidebar", sidebar);
+    setSidebar((prev) => {
+      return !prev;
+    });
+  };
 
   return (
     <div className={backgroundClasses}>
@@ -40,7 +49,7 @@ const NavigationBase = (props: NavigationBaseProps) => {
                   props.leftArrangement || ""
                 )}
               >
-                {<props.left />}
+                {<props.left sidebar={sidebar} toggleSidebar={toggleSidebar} />}
               </div>
             )}
             {/* Middle */}
@@ -51,7 +60,12 @@ const NavigationBase = (props: NavigationBaseProps) => {
                   props.middleArrangement || ""
                 )}
               >
-                {<props.middle />}
+                {
+                  <props.middle
+                    sidebar={sidebar}
+                    toggleSidebar={toggleSidebar}
+                  />
+                }
               </div>
             )}
             {/* Right */}
@@ -62,11 +76,22 @@ const NavigationBase = (props: NavigationBaseProps) => {
                   props.rightArrangement || ""
                 )}
               >
-                {<props.right />}
+                {
+                  <props.right
+                    sidebar={sidebar}
+                    toggleSidebar={toggleSidebar}
+                  />
+                }
               </div>
             )}
           </div>
         </div>
+
+        {props.sidebar !== undefined && (
+          <>
+            {<props.sidebar sidebar={sidebar} toggleSidebar={toggleSidebar} />}
+          </>
+        )}
       </div>
     </div>
   );